move test dialogues in separate module
