+++ tests/shared/core/test_dialogues.py
@@ -5,15 +5,15 @@ import pytest
 from rasa.shared.core.conversation import Dialogue
 from rasa.shared.core.domain import Domain
 from rasa.core.tracker_store import InMemoryTrackerStore
-from tests.conftest import (
+from tests.dialogues import (
     TEST_DIALOGUES,
     TEST_DEFAULT_DIALOGUE,
-    EXAMPLE_DOMAINS,
+    TEST_DOMAINS_FOR_DIALOGUES,
 )
 from tests.core.utilities import tracker_from_dialogue
 
 
-@pytest.mark.parametrize("pair", zip(TEST_DIALOGUES, EXAMPLE_DOMAINS))
+@pytest.mark.parametrize("pair", zip(TEST_DIALOGUES, TEST_DOMAINS_FOR_DIALOGUES))
 def test_inmemory_tracker_store(pair):
     dialogue, domainpath = pair
     domain = Domain.load(domainpath)
